+++ src/main/java/gov/cabinetoffice/gap/adminbackend/services/SchemeEditorService.java
@@ -3,6 +3,10 @@ package gov.cabinetoffice.gap.adminbackend.services;
 import gov.cabinetoffice.gap.adminbackend.config.UserServiceConfig;
 import gov.cabinetoffice.gap.adminbackend.dtos.schemes.SchemeEditorsDTO;
 import gov.cabinetoffice.gap.adminbackend.dtos.user.UserEmailRequestDto;
+import gov.cabinetoffice.gap.adminbackend.exceptions.FieldViolationException;
+import gov.cabinetoffice.gap.adminbackend.models.AdminSession;
+import gov.cabinetoffice.gap.adminbackend.utils.HelperUtils;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.reactive.function.BodyInserters;
 import gov.cabinetoffice.gap.adminbackend.dtos.user.UserEmailResponseDto;
 import gov.cabinetoffice.gap.adminbackend.entities.GrantAdmin;
@@ -15,6 +19,7 @@ import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
@@ -24,6 +29,7 @@ import java.util.Objects;
 public class SchemeEditorService {
     private final SchemeService schemeService;
     private final AwsEncryptionServiceImpl awsEncryptionService;
+    private final UserService userService;
     private final SchemeRepository schemeRepo;
     private final UserServiceConfig userServiceConfig;
     private final WebClient.Builder webClientBuilder;
@@ -79,6 +85,31 @@ public class SchemeEditorService {
 
         return Objects.requireNonNull(response).stream().map((item) ->
                 awsEncryptionService.decryptField(item.emailAddress())).toList();
+    }
+
+    public SchemeEntity addEditorToScheme(Integer schemeId, String editorEmailAddress, String jwt) {
+        AdminSession session = HelperUtils.getAdminSessionForAuthenticatedUser();
+        SchemeEntity scheme = this.schemeRepo.findById(schemeId).orElseThrow(EntityNotFoundException::new);
+        List<GrantAdmin> existingEditors = scheme.getGrantAdmins();
+        GrantAdmin editorToAdd;
+        try {
+            editorToAdd = userService.getGrantAdminIdFromUserServiceEmail(editorEmailAddress, jwt);
+        } catch (Exception e) {
+            throw new FieldViolationException("editorEmailAddress", "This account does not have an 'Administrator' account.");
+        }
+
+        if (existingEditors.stream().anyMatch(editor -> editor.getId().equals(editorToAdd.getId()))) {
+            throw new FieldViolationException("editorEmailAddress", "This email address is already an editor for this scheme");
+        }
+
+        if (!scheme.getCreatedBy().equals(session.getGrantAdminId())) {
+            throw new AccessDeniedException(
+                    "You are unable to add an editor to the scheme with id " + schemeId);
+        }
+
+        scheme.addAdmin(editorToAdd);
+        this.schemeRepo.save(scheme);
+        return scheme;
     }
 
 }
