+++ src/test/java/gov/cabinetoffice/gap/adminbackend/services/SchemeEditorServiceTest.java
@@ -8,9 +8,13 @@ import gov.cabinetoffice.gap.adminbackend.entities.GapUser;
 import gov.cabinetoffice.gap.adminbackend.entities.GrantAdmin;
 import gov.cabinetoffice.gap.adminbackend.entities.SchemeEntity;
 import gov.cabinetoffice.gap.adminbackend.enums.SchemeEditorRoleEnum;
+import gov.cabinetoffice.gap.adminbackend.exceptions.FieldViolationException;
 import gov.cabinetoffice.gap.adminbackend.exceptions.SchemeEntityException;
+import gov.cabinetoffice.gap.adminbackend.repositories.GrantAdminRepository;
 import gov.cabinetoffice.gap.adminbackend.repositories.SchemeRepository;
 import gov.cabinetoffice.gap.adminbackend.services.encryption.AwsEncryptionServiceImpl;
+import gov.cabinetoffice.gap.adminbackend.testdata.generators.RandomSchemeGenerator;
+import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -22,6 +26,8 @@ import org.springframework.web.client.RestClientException;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 
+import javax.persistence.EntityNotFoundException;
+
 import static org.mockito.Mockito.*;
 
 import java.util.*;
@@ -32,7 +38,6 @@ public class SchemeEditorServiceTest {
     @Mock
     private WebClient.Builder webClientBuilder;
 
-
     @Mock
     private SchemeService schemeService;
 
@@ -41,6 +46,12 @@ public class SchemeEditorServiceTest {
 
     @Mock
     private SchemeRepository schemeRepository;
+
+    @Mock
+    private UserService userService;
+
+    @Mock
+    private GrantAdminRepository grantAdminRepository;
 
     @Mock
     private AwsEncryptionServiceImpl awsEncryptionService;
@@ -164,4 +175,53 @@ public class SchemeEditorServiceTest {
             schemeEditorService.getEditorsFromSchemeId(1, "authHeader");
         });
     }
+
+    @Test
+    void addEditorToSchemeReturnsSchemeWithGrantAdmin() {
+        final SchemeEntity testScheme = RandomSchemeGenerator.randomSchemeEntity().build();
+        final GrantAdmin testAdmin = GrantAdmin.builder().id(1).build();
+        final SchemeEntity patchedScheme = RandomSchemeGenerator.randomSchemeEntity().build();
+        patchedScheme.addAdmin(testAdmin);
+
+        Mockito.when(SchemeEditorServiceTest.this.userService.getGrantAdminIdFromUserServiceEmail("test@test.gov", "jwt"))
+                .thenReturn(testAdmin);
+        Mockito.when(SchemeEditorServiceTest.this.schemeRepository.findById(1)).thenReturn(Optional.of(testScheme));
+        Mockito.when(SchemeEditorServiceTest.this.grantAdminRepository.findById(1)).thenReturn(Optional.of(testAdmin));
+        Mockito.when(SchemeEditorServiceTest.this.schemeRepository.save(testScheme)).thenReturn(patchedScheme);
+
+        SchemeEditorServiceTest.this.schemeEditorService.addEditorToScheme(1, "test@test.gov", "jwt");
+        AssertionsForClassTypes.assertThat(patchedScheme.getGrantAdmins().contains(testAdmin)).isTrue();
+
+    }
+
+    @Test
+    void addEditorToSchemeThrowsAnErrorIfSchemeIsNotPresent() {
+        Mockito.when(SchemeEditorServiceTest.this.schemeRepository.findById(1)).thenReturn(Optional.empty());
+
+        AssertionsForClassTypes.assertThatThrownBy(
+                        () -> SchemeEditorServiceTest.this.schemeEditorService.addEditorToScheme(1, "test@test.gov", "jwt"))
+                .isInstanceOf(EntityNotFoundException.class);
+    }
+
+    @Test
+    void addEditorToSchemeThrowsAnErrorIfGrantAdminIsNotPresent() {
+        final SchemeEntity testScheme = RandomSchemeGenerator.randomSchemeEntity().build();
+        Mockito.when(SchemeEditorServiceTest.this.schemeRepository.findById(1)).thenReturn(Optional.of(testScheme));
+        Mockito.when(SchemeEditorServiceTest.this.userService.getGrantAdminIdFromUserServiceEmail("test@test.gov", "jwt"))
+                .thenThrow(new EntityNotFoundException());
+    }
+
+    @Test
+    void addEditorThrowsAnErrorIfGrantAdminIsAlreadyAnEditor() {
+        final SchemeEntity testScheme = RandomSchemeGenerator.randomSchemeEntity().build();
+        final GrantAdmin testAdmin = GrantAdmin.builder().id(1).build();
+        testScheme.addAdmin(testAdmin);
+        Mockito.when(SchemeEditorServiceTest.this.schemeRepository.findById(1)).thenReturn(Optional.of(testScheme));
+        Mockito.when(SchemeEditorServiceTest.this.userService.getGrantAdminIdFromUserServiceEmail("test@test.gov", "jwt"))
+                .thenReturn(testAdmin);
+        AssertionsForClassTypes.assertThatThrownBy(
+                        () -> SchemeEditorServiceTest.this.schemeEditorService.addEditorToScheme(1, "test@test.gov", "jwt"))
+                .isInstanceOf(FieldViolationException.class);
+    }
+
 }
