+++ src/test/java/gov/cabinetoffice/gap/adminbackend/services/SchemeServiceTest.java
@@ -391,53 +391,4 @@ class SchemeServiceTest {
                 .isInstanceOf(SchemeEntityException.class).hasMessage(
                         "Update grant ownership failed: Something went wrong while trying to find scheme with id: 1");
     }
-
-    @Test
-    void addEditorToSchemeReturnsSchemeWithGrantAdmin() {
-        final SchemeEntity testScheme = RandomSchemeGenerator.randomSchemeEntity().build();
-        final GrantAdmin testAdmin = GrantAdmin.builder().id(1).build();
-        final SchemeEntity patchedScheme = RandomSchemeGenerator.randomSchemeEntity().build();
-        patchedScheme.addAdmin(testAdmin);
-
-        Mockito.when(SchemeServiceTest.this.userService.getGrantAdminIdFromUserServiceEmail("test@test.gov", "jwt"))
-                .thenReturn(testAdmin);
-        Mockito.when(SchemeServiceTest.this.schemeRepository.findById(1)).thenReturn(Optional.of(testScheme));
-        Mockito.when(SchemeServiceTest.this.grantAdminRepository.findById(1)).thenReturn(Optional.of(testAdmin));
-        Mockito.when(SchemeServiceTest.this.schemeRepository.save(testScheme)).thenReturn(patchedScheme);
-
-        SchemeServiceTest.this.schemeService.addEditorToScheme(1, "test@test.gov", "jwt");
-        AssertionsForClassTypes.assertThat(patchedScheme.getGrantAdmins().contains(testAdmin)).isTrue();
-
-    }
-
-    @Test
-    void addEditorToSchemeThrowsAnErrorIfSchemeIsNotPresent() {
-        Mockito.when(SchemeServiceTest.this.schemeRepository.findById(1)).thenReturn(Optional.empty());
-
-        AssertionsForClassTypes.assertThatThrownBy(
-                        () -> SchemeServiceTest.this.schemeService.addEditorToScheme(1, "test@test.gov", "jwt"))
-                .isInstanceOf(EntityNotFoundException.class);
-    }
-
-    @Test
-    void addEditorToSchemeThrowsAnErrorIfGrantAdminIsNotPresent() {
-        final SchemeEntity testScheme = RandomSchemeGenerator.randomSchemeEntity().build();
-        Mockito.when(SchemeServiceTest.this.schemeRepository.findById(1)).thenReturn(Optional.of(testScheme));
-        Mockito.when(SchemeServiceTest.this.userService.getGrantAdminIdFromUserServiceEmail("test@test.gov", "jwt"))
-                .thenThrow(new EntityNotFoundException());
-    }
-
-    @Test
-    void addEditorThrowsAnErrorIfGrantAdminIsAlreadyAnEditor() {
-        final SchemeEntity testScheme = RandomSchemeGenerator.randomSchemeEntity().build();
-        final GrantAdmin testAdmin = GrantAdmin.builder().id(1).build();
-        testScheme.addAdmin(testAdmin);
-        Mockito.when(SchemeServiceTest.this.schemeRepository.findById(1)).thenReturn(Optional.of(testScheme));
-        Mockito.when(SchemeServiceTest.this.userService.getGrantAdminIdFromUserServiceEmail("test@test.gov", "jwt"))
-                .thenReturn(testAdmin);
-        AssertionsForClassTypes.assertThatThrownBy(
-                        () -> SchemeServiceTest.this.schemeService.addEditorToScheme(1, "test@test.gov", "jwt"))
-                .isInstanceOf(FieldViolationException.class);
-    }
-
 }
